FEATURE: UART rx routine will receive one char from now on instead of receiving one frame at a time.
RATIONALE: when receiving a frame the end of frame is unknown and hence
the rx routine waits for 'x' milli seconds, before concluding a frame
done. But if another packet is received by Xbee within that window that
frame will be appended to the previous one! Processing char by char is
easy both in PC and uC context!